Split the largest generator methods into dedicated partial files
PInvokeGenerator.VisitDecl.cs and PInvokeGenerator.cs were dominated by a few very large methods: VisitRecordDecl (~1963 lines, ~47% of VisitDecl.cs), Close (~1444 lines, ~37% of PInvokeGenerator.cs), and VisitVarDecl (~290 lines). Their many nested local functions are the sole users of the enclosing method state, so a local function is the correct expression of that coupling; the readability problem is the sheer size of the host files, not the nesting.

Move each method verbatim (nested locals intact) into its own partial file -- PInvokeGenerator.VisitRecordDecl.cs, PInvokeGenerator.VisitVarDecl.cs, and PInvokeGenerator.Close.cs -- using the minimal public partial class PInvokeGenerator declaration matching the other partial files. No code within the methods changed; generated output is byte-identical and all 3708 golden tests pass.
