+++ sources/ClangSharp.PInvokeGenerator/PInvokeGenerator.VisitDecl.cs
@@ -1356,1970 +1356,6 @@ public partial class PInvokeGenerator
         }
     }
 
-    private void VisitRecordDecl(RecordDecl recordDecl)
-    {
-        if (recordDecl.IsInjectedClassName)
-        {
-            // We shouldn't process injected records
-            return;
-        }
-
-        var nativeName = GetCursorName(recordDecl);
-        var name = GetRemappedCursorName(recordDecl);
-        var escapedName = EscapeName(name);
-
-        StartUsingOutputBuilder(name, includeTestOutput: true);
-        {
-            var cxxRecordDecl = recordDecl as CXXRecordDecl;
-            _cxxRecordDeclContext = cxxRecordDecl;
-
-            var hasVtbl = false;
-            var hasBaseVtbl = false;
-
-            if (cxxRecordDecl is not null)
-            {
-                hasVtbl = HasVtbl(cxxRecordDecl, out hasBaseVtbl);
-            }
-
-            var alignment = Math.Max(recordDecl.TypeForDecl.Handle.AlignOf, 1);
-            var maxAlignm = recordDecl.Fields.Any() ? recordDecl.Fields.Max((fieldDecl) => Math.Max(fieldDecl.Type.Handle.AlignOf, 1)) : alignment;
-
-            var isTopLevelStruct = _config.WithTypes.TryGetValue(name, out var withType) && withType.Equals("struct", StringComparison.Ordinal);
-            var generateTestsClass = !recordDecl.IsAnonymousStructOrUnion && recordDecl.DeclContext is not RecordDecl;
-            var testOutputStarted = false;
-
-            var nullableUuid = (Guid?)null;
-            var uuidName = "";
-
-            if (TryGetUuid(recordDecl, out var uuid))
-            {
-                nullableUuid = uuid;
-                uuidName = GetRemappedName($"IID_{nativeName}", recordDecl, tryRemapOperatorName: false, out var wasRemapped, skipUsing: true);
-
-                _uuidsToGenerate.Add(uuidName, uuid);
-
-                if ((_testOutputBuilder is not null) && (uuid != Guid.Empty))
-                {
-                    StartTestOutput(ref testOutputStarted, generateTestsClass, isTopLevelStruct);
-
-                    var className = GetClass(uuidName);
-
-                    _testOutputBuilder.AddUsingDirective("System");
-
-                    if (_config.DontUseUsingStaticsForGuidMember)
-                    {
-                        _testOutputBuilder.AddUsingDirective($"{GetNamespace(className)}");
-                    }
-                    else
-                    {
-                        _testOutputBuilder.AddUsingDirective($"static {GetNamespace(className)}.{className}");
-                    }
-
-                    _testOutputBuilder.WriteIndented("/// <summary>Validates that the <see cref=\"Guid\" /> of the <see cref=\"");
-                    _testOutputBuilder.Write(escapedName);
-                    _testOutputBuilder.WriteLine("\" /> struct is correct.</summary>");
-
-                    WithTestAttribute();
-
-                    _testOutputBuilder.WriteIndentedLine("public static void GuidOfTest()");
-                    _testOutputBuilder.WriteBlockStart();
-
-                    if (_config.GenerateTestsNUnit)
-                    {
-                        _testOutputBuilder.WriteIndented("Assert.That");
-                    }
-                    else if (_config.GenerateTestsXUnit)
-                    {
-                        _testOutputBuilder.WriteIndented("Assert.Equal");
-                    }
-
-                    _testOutputBuilder.Write("(typeof(");
-                    _testOutputBuilder.Write(escapedName);
-                    _testOutputBuilder.Write(").GUID, ");
-
-                    if (_config.GenerateTestsNUnit)
-                    {
-                        _testOutputBuilder.Write("Is.EqualTo(");
-                    }
-
-                    var usableUuidName = uuidName;
-                    if (_config.DontUseUsingStaticsForGuidMember)
-                    {
-                        usableUuidName = $"{className}.{usableUuidName}";
-                    }
-
-                    _testOutputBuilder.Write(usableUuidName);
-
-                    if (_config.GenerateTestsNUnit)
-                    {
-                        _testOutputBuilder.Write(')');
-                    }
-
-                    _testOutputBuilder.Write(')');
-                    _testOutputBuilder.WriteSemicolon();
-                    _testOutputBuilder.WriteNewline();
-                    _testOutputBuilder.WriteBlockEnd();
-                    _testOutputBuilder.NeedsNewline = true;
-                }
-            }
-
-            var hasGuidMember = _config.GenerateGuidMember && !string.IsNullOrWhiteSpace(uuidName);
-
-            var layoutKind = recordDecl.IsUnion
-                ? LayoutKind.Explicit
-                : LayoutKind.Sequential;
-
-            long alignment32 = -1;
-            long alignment64 = -1;
-
-            GetTypeSize(recordDecl, recordDecl.TypeForDecl, ref alignment32, ref alignment64, out var size32, out var size64);
-
-            string[]? baseTypeNames = null;
-
-            string? nativeNameWithExtras = null, nativeInheritance = null;
-            if ((cxxRecordDecl is not null) && cxxRecordDecl.Bases.Any())
-            {
-                var nativeTypeNameBuilder = new StringBuilder();
-                var baseTypeNamesBuilder = new List<string>();
-
-                _ = nativeTypeNameBuilder.Append(recordDecl.IsUnion ? "union " : "struct ");
-                _ = nativeTypeNameBuilder.Append(nativeName);
-                _ = nativeTypeNameBuilder.Append(" : ");
-
-                var baseName = GetRemappedCursorName(cxxRecordDecl.Bases[0].Referenced, out var nativeBaseName, skipUsing: !_config.GenerateMarkerInterfaces);
-
-                baseTypeNamesBuilder.Add(baseName);
-                _ = nativeTypeNameBuilder.Append(nativeBaseName);
-
-                for (var i = 1; i < cxxRecordDecl.Bases.Count; i++)
-                {
-                    _ = nativeTypeNameBuilder.Append(", ");
-                    baseName = GetRemappedCursorName(cxxRecordDecl.Bases[i].Referenced, out nativeBaseName, skipUsing: !_config.GenerateMarkerInterfaces);
-
-                    baseTypeNamesBuilder.Add(baseName);
-                    _ = nativeTypeNameBuilder.Append(nativeBaseName);
-                }
-
-                nativeNameWithExtras = nativeTypeNameBuilder.ToString();
-                nativeInheritance = GetCursorName(cxxRecordDecl.Bases[cxxRecordDecl.Bases.Count - 1].Referenced);
-                baseTypeNames = [.. baseTypeNamesBuilder];
-            }
-
-            if (!TryGetRemappedValue(recordDecl, _config._withPackings, out var pack))
-            {
-                pack = alignment < maxAlignm ? alignment.ToString(CultureInfo.InvariantCulture) : null;
-            }
-
-            var desc = new StructDesc {
-                AccessSpecifier = GetAccessSpecifier(recordDecl, matchStar: true),
-                EscapedName = escapedName,
-                IsUnsafe = IsUnsafe(recordDecl) || hasGuidMember,
-                HasVtbl = hasVtbl || hasBaseVtbl,
-                IsUnion = recordDecl.IsUnion,
-                Layout = new() {
-                    Alignment32 = alignment32,
-                    Alignment64 = alignment64,
-                    Size32 = size32,
-                    Size64 = size64,
-                    Pack = pack,
-                    MaxFieldAlignment = maxAlignm,
-                    Kind = layoutKind
-                },
-                Uuid = nullableUuid,
-                NativeType = nativeNameWithExtras,
-                NativeInheritance = _config.GenerateNativeInheritanceAttribute ? nativeInheritance : null,
-                Location = recordDecl.Location,
-                IsNested = recordDecl.DeclContext is TagDecl,
-                WriteCustomAttrs = static context => {
-                    (var recordDecl, var generator) = ((RecordDecl, PInvokeGenerator))context;
-
-                    generator.WithAttributes(recordDecl);
-                    generator.WithUsings(recordDecl);
-                },
-                CustomAttrGeneratorData = (recordDecl, this),
-            };
-            Debug.Assert(_outputBuilder is not null);
-
-            if (!isTopLevelStruct)
-            {
-                _outputBuilder.BeginStruct(in desc);
-            }
-            else
-            {
-                if (!_topLevelClassAttributes.TryGetValue(name, out var withAttributes))
-                {
-                    withAttributes = [];
-                }
-
-                if (!_topLevelClassUsings.TryGetValue(name, out var withUsings))
-                {
-                    withUsings = new HashSet<string>(StringComparer.Ordinal);
-                }
-
-                if (desc.LayoutAttribute is not null)
-                {
-                    withAttributes.Add($"StructLayout(LayoutKind.{desc.LayoutAttribute.Value}{((desc.Layout.Pack is not null) ? $", Pack = {desc.Layout.Pack}" : "")})");
-                    _ = withUsings.Add("System.Runtime.InteropServices");
-                }
-
-                if (desc.Uuid.HasValue)
-                {
-                    withAttributes.Add($"Guid(\"{nullableUuid.GetValueOrDefault().ToString("D", CultureInfo.InvariantCulture).ToUpperInvariant()}\")");
-                    _ = withUsings.Add("System.Runtime.InteropServices");
-                }
-
-                var nativeTypeName = desc.NativeType;
-
-                if (nativeTypeName is not null)
-                {
-                    foreach (var entry in _config.NativeTypeNamesToStrip)
-                    {
-                        nativeTypeName = nativeTypeName.Replace(entry, "", StringComparison.Ordinal);
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(nativeTypeName))
-                    {
-                        withAttributes.Add($"NativeTypeName(\"{EscapeString(nativeTypeName)}\")");
-                        _ = withUsings.Add(GetNamespace("NativeTypeNameAttribute"));
-                    }
-                }
-
-                if (_config.GenerateNativeInheritanceAttribute && (desc.NativeInheritance is not null))
-                {
-                    withAttributes.Add($"NativeInheritance(\"{desc.NativeInheritance}\")");
-                    _ = withUsings.Add(GetNamespace("NativeInheritanceAttribute"));
-                }
-
-                if (_config.GenerateSourceLocationAttribute && (desc.Location is not null))
-                {
-                    desc.Location.Value.GetFileLocation(out var file, out var line, out var column, out _);
-                    withAttributes.Add($"SourceLocation(\"{EscapeString(file.Name.ToString())}\", {line}, {column})");
-                    _ = withUsings.Add(GetNamespace("SourceLocationAttribute"));
-                }
-
-                if (withAttributes.Count != 0)
-                {
-                    _topLevelClassAttributes[name] = withAttributes;
-                }
-
-                if (withUsings.Count != 0)
-                {
-                    _topLevelClassUsings[name] = withUsings;
-                }
-
-                if (desc.IsUnsafe)
-                {
-                    _topLevelClassIsUnsafe[name] = true;
-                }
-
-                if (hasGuidMember)
-                {
-                    _topLevelClassHasGuidMember[name] = true;
-                }
-            }
-
-            if (hasGuidMember)
-            {
-                var valueDesc = new ValueDesc {
-                    AccessSpecifier = AccessSpecifier.None,
-                    TypeName = "Guid*",
-                    EscapedName = "INativeGuid.NativeGuid",
-                    ParentName = name,
-                    Kind = ValueKind.GuidMember,
-                    Flags = ValueFlags.Initializer,
-                };
-
-                var uuidClassName = GetClass(uuidName);
-
-                _outputBuilder.EmitUsingDirective("System");
-                _outputBuilder.EmitUsingDirective("System.Runtime.CompilerServices");
-
-                if (_config.DontUseUsingStaticsForGuidMember)
-                {
-                    _outputBuilder.EmitUsingDirective($"{GetNamespace(uuidClassName)}");
-                }
-                else
-                {
-                    _outputBuilder.EmitUsingDirective($"static {GetNamespace(uuidClassName)}.{uuidClassName}");
-                }
-
-                var usableUuidName = uuidName;
-                if (_config.DontUseUsingStaticsForGuidMember)
-                {
-                    usableUuidName = $"{uuidClassName}.{usableUuidName}";
-                }
-
-                _outputBuilder.BeginValue(in valueDesc);
-
-                var code = _outputBuilder.BeginCSharpCode();
-
-                code.Write("(Guid*)Unsafe.AsPointer(");
-
-                if (!_config.GenerateLatestCode)
-                {
-                    code.Write("ref Unsafe.AsRef(");
-                }
-
-                code.Write("in ");
-                code.Write(usableUuidName);
-                code.Write(')');
-
-                if (!_config.GenerateLatestCode)
-                {
-                    code.Write(')');
-                }
-
-                _outputBuilder.EndCSharpCode(code);
-
-                _outputBuilder.EndValue(in valueDesc);
-            }
-
-            if (hasVtbl || (hasBaseVtbl && (cxxRecordDecl is not null) && !HasBaseField(cxxRecordDecl)))
-            {
-                var fieldDesc = new FieldDesc {
-                    AccessSpecifier = AccessSpecifier.Public,
-                    NativeTypeName = null,
-                    EscapedName = "lpVtbl",
-                    Offset = null,
-                    NeedsNewKeyword = false,
-                };
-                _outputBuilder.BeginField(in fieldDesc);
-
-                if (_config.GenerateExplicitVtbls)
-                {
-                    if (_config.GenerateMarkerInterfaces && !_config.GenerateCompatibleCode)
-                    {
-                        _outputBuilder.WriteRegularField($"Vtbl<{nativeName}>*", "lpVtbl");
-                    }
-                    else
-                    {
-                        _outputBuilder.WriteRegularField("Vtbl*", "lpVtbl");
-                    }
-                }
-                else
-                {
-                    _outputBuilder.WriteRegularField("void**", "lpVtbl");
-                }
-
-                _outputBuilder.EndField(in fieldDesc);
-            }
-
-            if (cxxRecordDecl is not null)
-            {
-                for (var index = 0; index < cxxRecordDecl.Bases.Count; index++)
-                {
-                    var cxxBaseSpecifier = cxxRecordDecl.Bases[index];
-                    var baseCxxRecordDecl = GetRecordDecl(cxxBaseSpecifier);
-
-                    if (HasField(baseCxxRecordDecl) && !IsBaseExcluded(cxxRecordDecl, baseCxxRecordDecl, cxxBaseSpecifier, out var baseFieldName))
-                    {
-                        var parent = GetRemappedCursorName(baseCxxRecordDecl);
-
-                        var fieldDesc = new FieldDesc {
-                            AccessSpecifier = GetAccessSpecifier(baseCxxRecordDecl, matchStar: true),
-                            NativeTypeName = null,
-                            EscapedName = baseFieldName,
-                            Offset = null,
-                            NeedsNewKeyword = false,
-                            InheritedFrom = parent,
-                            Location = cxxBaseSpecifier.Location,
-                        };
-
-                        _outputBuilder.BeginField(in fieldDesc);
-                        _outputBuilder.WriteRegularField(parent, baseFieldName);
-                        _outputBuilder.EndField(in fieldDesc);
-                    }
-                }
-            }
-
-            if ((_testOutputBuilder is not null) && generateTestsClass && !_config.GenerateDisableRuntimeMarshalling)
-            {
-                StartTestOutput(ref testOutputStarted, generateTestsClass, isTopLevelStruct);
-
-                _testOutputBuilder.WriteIndented("/// <summary>Validates that the <see cref=\"");
-                _testOutputBuilder.Write(escapedName);
-                _testOutputBuilder.WriteLine("\" /> struct is blittable.</summary>");
-
-                WithTestAttribute();
-
-                _testOutputBuilder.WriteIndentedLine("public static void IsBlittableTest()");
-                _testOutputBuilder.WriteBlockStart();
-
-                WithTestAssertEqual($"sizeof({escapedName})", $"Marshal.SizeOf<{escapedName}>()");
-
-                _testOutputBuilder.WriteBlockEnd();
-                _testOutputBuilder.NeedsNewline = true;
-
-                _testOutputBuilder.WriteIndented("/// <summary>Validates that the <see cref=\"");
-                _testOutputBuilder.Write(escapedName);
-                _testOutputBuilder.WriteLine("\" /> struct has the right <see cref=\"LayoutKind\" />.</summary>");
-
-                WithTestAttribute();
-
-                _testOutputBuilder.WriteIndented("public static void IsLayout");
-
-                if (recordDecl.IsUnion)
-                {
-                    _testOutputBuilder.Write("Explicit");
-                }
-                else
-                {
-                    _testOutputBuilder.Write("Sequential");
-                }
-
-                _testOutputBuilder.WriteLine("Test()");
-                _testOutputBuilder.WriteBlockStart();
-
-                WithTestAssertTrue($"typeof({escapedName}).Is{(recordDecl.IsUnion ? "ExplicitLayout" : "LayoutSequential")}");
-
-                _testOutputBuilder.WriteBlockEnd();
-                _testOutputBuilder.NeedsNewline = true;
-
-                if ((size32 == 0 || size64 == 0) && !TryGetUuid(recordDecl, out _))
-                {
-                    AddDiagnostic(DiagnosticLevel.Info, $"{escapedName} has a size of 0", recordDecl);
-                }
-
-                _testOutputBuilder.WriteIndented("/// <summary>Validates that the <see cref=\"");
-                _testOutputBuilder.Write(escapedName);
-                _testOutputBuilder.WriteLine("\" /> struct has the correct size.</summary>");
-
-                WithTestAttribute();
-
-                _testOutputBuilder.WriteIndentedLine("public static void SizeOfTest()");
-                _testOutputBuilder.WriteBlockStart();
-
-                if (size32 != size64)
-                {
-                    _testOutputBuilder.AddUsingDirective("System");
-
-                    _testOutputBuilder.WriteIndentedLine("if (Environment.Is64BitProcess)");
-                    _testOutputBuilder.WriteBlockStart();
-
-                    WithTestAssertEqual($"{Math.Max(size64, 1)}", $"sizeof({escapedName})");
-
-                    _testOutputBuilder.WriteBlockEnd();
-                    _testOutputBuilder.WriteIndentedLine("else");
-                    _testOutputBuilder.WriteBlockStart();
-                }
-
-                WithTestAssertEqual($"{Math.Max(size32, 1)}", $"sizeof({escapedName})");
-
-                if (size32 != size64)
-                {
-                    _testOutputBuilder.WriteBlockEnd();
-                }
-
-                _testOutputBuilder.WriteBlockEnd();
-            }
-
-            var bitfieldDescs = GetBitfieldDescs(recordDecl);
-            var bitfieldIndex = (bitfieldDescs.Length == 1) ? -1 : 0;
-
-            var bitfieldPreviousSize = 0L;
-            var bitfieldRemainingBits = 0L;
-
-            foreach (var fieldDecl in recordDecl.Fields)
-            {
-                if (fieldDecl.IsBitField)
-                {
-                    VisitBitfieldDecl(fieldDecl, bitfieldDescs, recordDecl, contextName: "", ref bitfieldIndex, ref bitfieldPreviousSize, ref bitfieldRemainingBits);
-                }
-                else
-                {
-                    bitfieldPreviousSize = 0;
-                    bitfieldRemainingBits = 0;
-                }
-
-                Visit(fieldDecl);
-                _outputBuilder.WriteDivider();
-            }
-
-            Visit(recordDecl.IndirectFields);
-
-            if (cxxRecordDecl is not null)
-            {
-                foreach (var cxxConstructorDecl in cxxRecordDecl.Ctors)
-                {
-                    if (!IsExcluded(cxxConstructorDecl))
-                    {
-                        Visit(cxxConstructorDecl);
-                        _outputBuilder.WriteDivider();
-                    }
-                }
-
-                if (cxxRecordDecl.HasUserDeclaredDestructor)
-                {
-                    var cxxDestructorDecl = cxxRecordDecl.Destructor;
-
-                    if (cxxDestructorDecl == null)
-                    {
-                        AddDiagnostic(DiagnosticLevel.Warning, "Record has user declared destructor, but Destructor property was null. Generated bindings may be incomplete.", cxxRecordDecl);
-                    }
-                    else if (!cxxDestructorDecl.IsVirtual && !IsExcluded(cxxDestructorDecl))
-                    {
-                        Visit(cxxDestructorDecl);
-                        _outputBuilder.WriteDivider();
-                    }
-                }
-
-                if (hasVtbl || hasBaseVtbl)
-                {
-                    OutputDelegateSignatures(cxxRecordDecl, cxxRecordDecl);
-                }
-            }
-
-            var excludedCursors = recordDecl.Fields.AsEnumerable<Cursor>().Concat(recordDecl.IndirectFields);
-
-            if (cxxRecordDecl is not null)
-            {
-                OutputMethods(cxxRecordDecl, cxxRecordDecl);
-                excludedCursors = excludedCursors.Concat(cxxRecordDecl.Methods);
-            }
-
-            Visit(recordDecl.Decls, excludedCursors);
-
-            foreach (var array in recordDecl.Fields.Where(IsTypeConstantOrIncompleteArray))
-            {
-                VisitConstantOrIncompleteArrayFieldDecl(recordDecl, array);
-            }
-
-            if (hasVtbl || hasBaseVtbl)
-            {
-                Debug.Assert(cxxRecordDecl is not null);
-
-                if (!_config.GenerateCompatibleCode)
-                {
-                    _outputBuilder.EmitCompatibleCodeSupport();
-                }
-
-                if (_config.ExcludeFnptrCodegen)
-                {
-                    _outputBuilder.EmitFnPtrSupport();
-                }
-
-                OutputVtblHelperMethods(cxxRecordDecl, cxxRecordDecl);
-
-                if (_config.GenerateMarkerInterfaces)
-                {
-                    if (_outputBuilder is CSharpOutputBuilder csharpOutputBuilder)
-                    {
-                        csharpOutputBuilder.NeedsNewline = true;
-                    }
-
-                    _outputBuilder.BeginMarkerInterface(baseTypeNames);
-                    OutputMarkerInterfaces(cxxRecordDecl, cxxRecordDecl);
-                    _outputBuilder.EndMarkerInterface();
-                }
-
-                if (_config.GenerateExplicitVtbls || _config.GenerateTrimmableVtbls)
-                {
-                    if (_outputBuilder is CSharpOutputBuilder csharpOutputBuilder)
-                    {
-                        csharpOutputBuilder.NeedsNewline = true;
-                    }
-
-                    _outputBuilder.BeginExplicitVtbl();
-                    OutputVtblEntries(cxxRecordDecl, cxxRecordDecl);
-                    _outputBuilder.EndExplicitVtbl();
-                }
-            }
-
-            if (!isTopLevelStruct)
-            {
-                _outputBuilder.EndStruct(in desc);
-
-                if ((_testOutputBuilder is not null) && generateTestsClass && testOutputStarted)
-                {
-                    _testOutputBuilder.WriteBlockEnd();
-                }
-            }
-
-            _cxxRecordDeclContext = null;
-        }
-        StopUsingOutputBuilder();
-
-        string FixupNameForMultipleHits(CXXMethodDecl cxxMethodDecl)
-        {
-            var remappedName = GetRemappedCursorName(cxxMethodDecl);
-            var overloadIndex = GetOverloadIndex(cxxMethodDecl);
-
-            if (overloadIndex != 0)
-            {
-                remappedName = $"{remappedName}{overloadIndex}";
-            }
-            return remappedName;
-        }
-
-        void OutputDelegateSignatures(CXXRecordDecl rootCxxRecordDecl, CXXRecordDecl cxxRecordDecl)
-        {
-            if (!_config.ExcludeFnptrCodegen)
-            {
-                return;
-            }
-
-            foreach (var cxxBaseSpecifier in cxxRecordDecl.Bases)
-            {
-                var baseCxxRecordDecl = GetRecordDecl(cxxBaseSpecifier);
-                OutputDelegateSignatures(rootCxxRecordDecl, baseCxxRecordDecl);
-            }
-
-            var cxxMethodDecls = cxxRecordDecl.Methods;
-
-            if (cxxMethodDecls.Count != 0)
-            {
-                foreach (var cxxMethodDecl in cxxMethodDecls.OrderBy((cxxmd) => cxxmd.VtblIndex))
-                {
-                    if (!cxxMethodDecl.IsVirtual)
-                    {
-                        continue;
-                    }
-
-                    if (IsExcluded(cxxMethodDecl, out var isExcludedByConflictingDefinition))
-                    {
-                        continue;
-                    }
-
-                    _outputBuilder.WriteDivider();
-
-                    var remappedName = FixupNameForMultipleHits(cxxMethodDecl);
-                    Debug.Assert(CurrentContext.Cursor == rootCxxRecordDecl);
-                    Visit(cxxMethodDecl);
-                }
-            }
-        }
-
-        void OutputMethods(CXXRecordDecl rootCxxRecordDecl, CXXRecordDecl cxxRecordDecl)
-        {
-            foreach (var cxxBaseSpecifier in cxxRecordDecl.Bases)
-            {
-                var baseCxxRecordDecl = GetRecordDecl(cxxBaseSpecifier);
-                OutputMethods(rootCxxRecordDecl, baseCxxRecordDecl);
-            }
-
-            var cxxMethodDecls = cxxRecordDecl.Methods;
-
-            if (cxxMethodDecls.Count != 0)
-            {
-                foreach (var cxxMethodDecl in cxxMethodDecls.OrderBy((cxxmd) => cxxmd.VtblIndex))
-                {
-                    if (cxxMethodDecl.IsVirtual)
-                    {
-                        continue;
-                    }
-
-                    if (cxxMethodDecl is CXXConstructorDecl or CXXDestructorDecl)
-                    {
-                        continue;
-                    }
-
-                    Debug.Assert(CurrentContext.Cursor == rootCxxRecordDecl);
-                    Visit(cxxMethodDecl);
-                    _outputBuilder.WriteDivider();
-                }
-            }
-        }
-
-        void OutputMarkerInterface(CXXRecordDecl cxxRecordDecl, CXXMethodDecl cxxMethodDecl)
-        {
-            if (!cxxMethodDecl.IsVirtual)
-            {
-                return;
-            }
-
-            if (IsExcluded(cxxMethodDecl, out var isExcludedByConflictingDefinition))
-            {
-                return;
-            }
-
-            if (_config.GenerateTrimmableVtbls && cxxMethodDecl.Parameters.Any((parmVarDecl) => IsType<PointerType>(parmVarDecl, parmVarDecl.Type, out var pointerType) &&
-                                                                                                IsType<FunctionType>(parmVarDecl, pointerType.PointeeType, out _)))
-            {
-                // This breaks trimming right now
-                return;
-            }
-
-            var currentContext = _context.AddLast((cxxMethodDecl, null));
-
-            var returnType = cxxMethodDecl.ReturnType;
-            var returnTypeName = GetRemappedTypeName(cxxMethodDecl, cxxRecordDecl, returnType, out var nativeTypeName);
-
-            var remappedName = FixupNameForMultipleHits(cxxMethodDecl);
-            var name = GetRemappedCursorName(cxxMethodDecl);
-            var needsReturnFixup = false;
-            var needsCastToTransparentStruct = false;
-
-            if (!IsTypeVoid(cxxMethodDecl, returnType))
-            {
-                needsReturnFixup = NeedsReturnFixup(cxxMethodDecl);
-                needsCastToTransparentStruct = _config.WithTransparentStructs.TryGetValue(returnTypeName, out var transparentStruct) && IsTransparentStructHandle(transparentStruct.Kind);
-            }
-
-            var desc = new FunctionOrDelegateDesc {
-                AccessSpecifier = AccessSpecifier.Public,
-                EscapedName = EscapeAndStripMethodName(name),
-                IsMemberFunction = true,
-                NativeTypeName = nativeTypeName,
-                HasFnPtrCodeGen = !_config.ExcludeFnptrCodegen,
-                IsCtxCxxRecord = true,
-                IsCxxRecordCtxUnsafe = IsUnsafe(cxxRecordDecl),
-                IsUnsafe = true,
-                NeedsReturnFixup = needsReturnFixup,
-                ReturnType = returnTypeName,
-                VtblIndex = _config.GenerateVtblIndexAttribute ? cxxMethodDecl.VtblIndex : -1,
-                Location = cxxMethodDecl.Location,
-                WriteCustomAttrs = static context => {
-                    (var cxxMethodDecl, var generator) = ((CXXMethodDecl, PInvokeGenerator))context;
-
-                    generator.WithAttributes(cxxMethodDecl);
-                    generator.WithUsings(cxxMethodDecl);
-                },
-                CustomAttrGeneratorData = (cxxMethodDecl, this),
-            };
-
-            var isUnsafe = true;
-            _outputBuilder.BeginFunctionOrDelegate(in desc, ref isUnsafe);
-
-            _outputBuilder.BeginFunctionInnerPrototype(in desc);
-
-            Visit(cxxMethodDecl.Parameters);
-
-            _outputBuilder.EndFunctionInnerPrototype(in desc);
-            _outputBuilder.EndFunctionOrDelegate(in desc);
-
-            Debug.Assert(_context.Last == currentContext);
-            _context.RemoveLast();
-        }
-
-        void OutputMarkerInterfaces(CXXRecordDecl rootCxxRecordDecl, CXXRecordDecl cxxRecordDecl)
-        {
-            // Don't process the base types because that's exposed via inheritance instead
-            var cxxMethodDecls = cxxRecordDecl.Methods;
-
-            if (cxxMethodDecls.Count != 0)
-            {
-                foreach (var cxxMethodDecl in cxxMethodDecls.OrderBy((cxxmd) => cxxmd.VtblIndex))
-                {
-                    OutputMarkerInterface(rootCxxRecordDecl, cxxMethodDecl);
-                }
-            }
-        }
-
-        void OutputVtblEntries(CXXRecordDecl rootCxxRecordDecl, CXXRecordDecl cxxRecordDecl)
-        {
-            foreach (var cxxBaseSpecifier in cxxRecordDecl.Bases)
-            {
-                var baseCxxRecordDecl = GetRecordDecl(cxxBaseSpecifier);
-                OutputVtblEntries(rootCxxRecordDecl, baseCxxRecordDecl);
-            }
-
-            var cxxMethodDecls = cxxRecordDecl.Methods;
-
-            if (cxxMethodDecls.Count != 0)
-            {
-                foreach (var cxxMethodDecl in cxxMethodDecls.OrderBy((cxxmd) => cxxmd.VtblIndex))
-                {
-                    OutputVtblEntry(rootCxxRecordDecl, cxxMethodDecl);
-                }
-            }
-        }
-
-        void OutputVtblEntry(CXXRecordDecl cxxRecordDecl, CXXMethodDecl cxxMethodDecl)
-        {
-            if (!cxxMethodDecl.IsVirtual)
-            {
-                return;
-            }
-
-            if (IsExcluded(cxxMethodDecl, out var isExcludedByConflictingDefinition))
-            {
-                return;
-            }
-
-            var cxxMethodDeclTypeName = GetRemappedTypeName(cxxMethodDecl, cxxRecordDecl, cxxMethodDecl.Type, out var nativeTypeName, skipUsing: false, ignoreTransparentStructsWhereRequired: true);
-
-            if (_config.GenerateMarkerInterfaces && !_config.ExcludeFnptrCodegen)
-            {
-                var cxxRecordDeclName = GetRemappedCursorName(cxxRecordDecl);
-                cxxMethodDeclTypeName = cxxMethodDeclTypeName.Replace($"<{cxxRecordDeclName}*,", "<TSelf*,", StringComparison.Ordinal);
-            }
-
-            var remappedName = FixupNameForMultipleHits(cxxMethodDecl);
-            var escapedName = EscapeAndStripMethodName(remappedName);
-
-            var desc = new FieldDesc {
-                AccessSpecifier = AccessSpecifier.Public,
-                NativeTypeName = nativeTypeName,
-                EscapedName = escapedName,
-                Offset = null,
-                NeedsNewKeyword = NeedsNewKeyword(remappedName),
-                Location = cxxMethodDecl.Location,
-                WriteCustomAttrs = static context => {
-                    (var cxxMethodDecl, var generator) = ((CXXMethodDecl, PInvokeGenerator))context;
-
-                    generator.WithAttributes(cxxMethodDecl);
-                    generator.WithUsings(cxxMethodDecl);
-                },
-                CustomAttrGeneratorData = (cxxMethodDecl, this),
-            };
-
-            _outputBuilder.BeginField(in desc);
-            _outputBuilder.WriteRegularField(cxxMethodDeclTypeName, escapedName);
-            _outputBuilder.EndField(in desc);
-
-            _outputBuilder.WriteDivider();
-        }
-
-        void OutputVtblHelperMethod(CXXRecordDecl cxxRecordDecl, CXXMethodDecl cxxMethodDecl)
-        {
-            if (!cxxMethodDecl.IsVirtual)
-            {
-                return;
-            }
-
-            if (IsExcluded(cxxMethodDecl, out var isExcludedByConflictingDefinition))
-            {
-                return;
-            }
-
-            var currentContext = _context.AddLast((cxxMethodDecl, null));
-
-            var returnType = cxxMethodDecl.ReturnType;
-            var returnTypeName = GetRemappedTypeName(cxxMethodDecl, cxxRecordDecl, returnType, out var nativeTypeName);
-
-            var remappedName = FixupNameForMultipleHits(cxxMethodDecl);
-            var name = GetRemappedCursorName(cxxMethodDecl);
-            var needsReturnFixup = false;
-            var needsCastToTransparentStruct = false;
-            var cxxRecordDeclName = GetRemappedCursorName(cxxRecordDecl);
-            var parentName = cxxRecordDeclName;
-            var isInherited = false;
-
-            var parent = cxxMethodDecl.Parent;
-            Debug.Assert(parent is not null);
-
-            if (parent != cxxRecordDecl)
-            {
-                parentName = GetRemappedCursorName(parent);
-                isInherited = true;
-            }
-
-            if (!IsTypeVoid(cxxMethodDecl, returnType))
-            {
-                needsReturnFixup = NeedsReturnFixup(cxxMethodDecl);
-                needsCastToTransparentStruct = _config.WithTransparentStructs.TryGetValue(returnTypeName, out var transparentStruct) && IsTransparentStructHandle(transparentStruct.Kind);
-            }
-
-            var desc = new FunctionOrDelegateDesc {
-                AccessSpecifier = AccessSpecifier.Public,
-                IsAggressivelyInlined = _config.GenerateAggressiveInlining,
-                EscapedName = EscapeAndStripMethodName(name),
-                ParentName = parentName,
-                IsMemberFunction = true,
-                IsInherited = isInherited,
-                NativeTypeName = nativeTypeName,
-                NeedsNewKeyword = NeedsNewKeyword(name, cxxMethodDecl.Parameters),
-                HasFnPtrCodeGen = !_config.ExcludeFnptrCodegen,
-                IsCtxCxxRecord = true,
-                IsCxxRecordCtxUnsafe = IsUnsafe(cxxRecordDecl),
-                IsReadOnly = IsReadonly(cxxMethodDecl),
-                IsUnsafe = true,
-                NeedsReturnFixup = needsReturnFixup,
-                ReturnType = returnTypeName,
-                VtblIndex = _config.GenerateVtblIndexAttribute ? cxxMethodDecl.VtblIndex : -1,
-                Location = cxxMethodDecl.Location,
-                HasBody = true,
-                WriteCustomAttrs = static context => {
-                    (var cxxMethodDecl, var generator) = ((CXXMethodDecl, PInvokeGenerator))context;
-
-                    generator.WithAttributes(cxxMethodDecl);
-                    generator.WithUsings(cxxMethodDecl);
-                },
-                CustomAttrGeneratorData = (cxxMethodDecl, this),
-            };
-
-            var isUnsafe = true;
-            _outputBuilder.BeginFunctionOrDelegate(in desc, ref isUnsafe);
-
-            _outputBuilder.BeginFunctionInnerPrototype(in desc);
-
-            Visit(cxxMethodDecl.Parameters);
-
-            _outputBuilder.EndFunctionInnerPrototype(in desc);
-            _outputBuilder.BeginBody();
-
-            var escapedCXXRecordDeclName = EscapeName(cxxRecordDeclName);
-
-            _outputBuilder.BeginInnerFunctionBody();
-            var body = _outputBuilder.BeginCSharpCode();
-
-            if (_config.GenerateCompatibleCode)
-            {
-                body.Write("fixed (");
-                body.Write(escapedCXXRecordDeclName);
-                body.WriteLine("* pThis = &this)");
-                body.WriteBlockStart();
-                body.WriteIndentation();
-            }
-
-            if (needsReturnFixup)
-            {
-                body.BeginMarker("fixup", new KeyValuePair<string, object>("type", "*result"));
-                body.Write(returnTypeName);
-                body.EndMarker("fixup");
-                body.Write(" result");
-                body.WriteSemicolon();
-                body.WriteNewline();
-                body.WriteIndentation();
-            }
-
-            if (!IsTypeVoid(cxxMethodDecl, returnType))
-            {
-                body.Write("return ");
-            }
-
-            if (needsCastToTransparentStruct)
-            {
-                body.Write("((");
-                body.Write(returnTypeName);
-                body.Write(")(");
-            }
-
-            if (needsReturnFixup)
-            {
-                body.Write('*');
-            }
-
-            if (_config.ExcludeFnptrCodegen)
-            {
-                body.Write("Marshal.GetDelegateForFunctionPointer<");
-                body.BeginMarker("delegate");
-                body.Write(PrefixAndStripMethodName(name, GetOverloadIndex(cxxMethodDecl)));
-                body.EndMarker("delegate");
-                body.Write(">(");
-            }
-
-            if (_config.GenerateExplicitVtbls)
-            {
-                body.Write("lpVtbl->");
-                body.BeginMarker("vtbl", new KeyValuePair<string, object>("explicit", true));
-                body.Write(EscapeAndStripMethodName(remappedName));
-                body.EndMarker("vtbl");
-            }
-            else
-            {
-                var cxxMethodDeclTypeName = GetRemappedTypeName(cxxMethodDecl, cxxRecordDecl, cxxMethodDecl.Type, out _, skipUsing: false, ignoreTransparentStructsWhereRequired: true);
-
-                if (!_config.ExcludeFnptrCodegen)
-                {
-                    body.Write('(');
-                }
-
-                body.Write('(');
-                body.Write(cxxMethodDeclTypeName);
-                body.Write(")(lpVtbl[");
-                body.BeginMarker("vtbl", new KeyValuePair<string, object>("explicit", false));
-                body.Write(cxxMethodDecl.VtblIndex);
-                body.EndMarker("vtbl");
-                body.Write("])");
-
-                if (!_config.ExcludeFnptrCodegen)
-                {
-                    body.Write(')');
-                }
-            }
-
-            if (_config.ExcludeFnptrCodegen)
-            {
-                body.Write(')');
-            }
-
-            body.Write('(');
-
-            body.BeginMarker("param", new KeyValuePair<string, object>("special", "thisPtr"));
-            if (_config.GenerateCompatibleCode)
-            {
-                body.Write("pThis");
-            }
-            else
-            {
-                body.Write('(');
-                body.Write(escapedCXXRecordDeclName);
-                body.Write("*)Unsafe.AsPointer(");
-
-                if (IsReadonly(cxxMethodDecl))
-                {
-                    if (!_config.GenerateLatestCode)
-                    {
-                        body.Write("ref Unsafe.AsRef(");
-                    }
-
-                    body.Write("in this");
-
-                    if (!_config.GenerateLatestCode)
-                    {
-                        body.Write(')');
-                    }
-                }
-                else
-                {
-                    body.Write("ref this");
-                }
-                body.Write(')');
-            }
-            body.EndMarker("param");
-
-            if (needsReturnFixup)
-            {
-                body.BeginMarker("param", new KeyValuePair<string, object>("special", "retFixup"));
-                body.Write(", &result");
-                body.EndMarker("param");
-            }
-
-            var parmVarDecls = cxxMethodDecl.Parameters;
-
-            for (var index = 0; index < parmVarDecls.Count; index++)
-            {
-                body.Write(", ");
-
-                var parmVarDeclName = GetRemappedCursorName(parmVarDecls[index]);
-                var escapedParmVarDeclName = EscapeName(parmVarDeclName);
-
-                if (parmVarDeclName.Equals("param", StringComparison.Ordinal))
-                {
-                    escapedParmVarDeclName += index;
-                }
-
-                body.BeginMarker("param", new KeyValuePair<string, object>("name", escapedParmVarDeclName));
-                body.Write(escapedParmVarDeclName);
-                body.EndMarker("param");
-            }
-
-            body.Write(')');
-
-            if (!_config.GenerateDisableRuntimeMarshalling && returnTypeName.Equals("bool", StringComparison.Ordinal))
-            {
-                // bool is not blittable when DisableRuntimeMarshalling is not specified, so we shouldn't use it for P/Invoke signatures
-                body.Write(" != 0");
-            }
-
-            if (needsCastToTransparentStruct)
-            {
-                body.Write("))");
-            }
-
-            body.WriteSemicolon();
-            body.WriteNewline();
-
-            if (_config.GenerateCompatibleCode)
-            {
-                body.WriteBlockEnd();
-            }
-
-            _outputBuilder.EndCSharpCode(body);
-            _outputBuilder.EndInnerFunctionBody();
-            _outputBuilder.EndBody();
-            _outputBuilder.EndFunctionOrDelegate(in desc);
-
-            Debug.Assert(_context.Last == currentContext);
-            _context.RemoveLast();
-        }
-
-        void OutputVtblHelperMethods(CXXRecordDecl rootCxxRecordDecl, CXXRecordDecl cxxRecordDecl)
-        {
-            foreach (var cxxBaseSpecifier in cxxRecordDecl.Bases)
-            {
-                var baseCxxRecordDecl = GetRecordDecl(cxxBaseSpecifier);
-                OutputVtblHelperMethods(rootCxxRecordDecl, baseCxxRecordDecl);
-            }
-
-            var cxxMethodDecls = cxxRecordDecl.Methods;
-
-            if (cxxMethodDecls.Count != 0)
-            {
-                foreach (var cxxMethodDecl in cxxMethodDecls.OrderBy((cxxmd) => cxxmd.VtblIndex))
-                {
-                    _outputBuilder.WriteDivider();
-                    OutputVtblHelperMethod(rootCxxRecordDecl, cxxMethodDecl);
-                }
-            }
-        }
-
-        void VisitBitfieldDecl(FieldDecl fieldDecl, BitfieldDesc[] bitfieldDescs, RecordDecl recordDecl, string contextName, ref int index, ref long previousSize, ref long remainingBits)
-        {
-            Debug.Assert(fieldDecl.IsBitField);
-
-            var typeName = GetRemappedTypeName(fieldDecl, context: null, fieldDecl.Type, out var nativeTypeName);
-
-            if (string.IsNullOrWhiteSpace(nativeTypeName))
-            {
-                nativeTypeName = typeName;
-            }
-
-            nativeTypeName += $" : {fieldDecl.BitWidthValue}";
-
-            var currentSize = fieldDecl.Type.Handle.SizeOf;
-            var bitfieldName = "_bitfield";
-
-            Type typeBacking;
-            string typeNameBacking;
-
-            var parent = fieldDecl.Parent;
-            Debug.Assert(parent is not null);
-
-            if ((!_config.GenerateUnixTypes && (currentSize != previousSize)) || (fieldDecl.BitWidthValue > remainingBits))
-            {
-                if (index >= 0)
-                {
-                    index++;
-                    bitfieldName += index.ToString(CultureInfo.InvariantCulture);
-                }
-
-                remainingBits = currentSize * 8;
-                previousSize = 0;
-
-                var bitfieldDesc = (index > 0) ? bitfieldDescs[index - 1] : bitfieldDescs[0];
-                typeBacking = bitfieldDesc.TypeBacking;
-                typeNameBacking = GetRemappedTypeName(fieldDecl, context: null, typeBacking, out _);
-
-                if (parent == recordDecl)
-                {
-                    var fieldDesc = new FieldDesc {
-                        AccessSpecifier = AccessSpecifier.Public,
-                        NativeTypeName = null,
-                        EscapedName = bitfieldName,
-                        Offset = parent.IsUnion ? 0 : null,
-                        NeedsNewKeyword = false,
-                        Location = fieldDecl.Location,
-                        WriteCustomAttrs = static context => {
-                            (var bitfieldDesc, var generator) = ((BitfieldDesc, PInvokeGenerator))context;
-
-                            if (!generator.Config.GenerateNativeBitfieldAttribute)
-                            {
-                                return;
-                            }
-
-                            var outputBuilder = generator._outputBuilder;
-                            Debug.Assert(outputBuilder is not null);
-
-                            foreach (var bitfieldRegion in bitfieldDesc.Regions)
-                            {
-                                outputBuilder.WriteCustomAttribute($"NativeBitfield(\"{bitfieldRegion.Name}\", offset: {bitfieldRegion.Offset}, length: {bitfieldRegion.Length})");
-                            }
-
-                            var namespaceName = generator.GetNamespace("NativeBitfieldAttribute");
-                            generator.AddUsingDirective(outputBuilder, namespaceName);
-                        },
-                        CustomAttrGeneratorData = (bitfieldDesc, this),
-                    };
-                    _outputBuilder.BeginField(in fieldDesc);
-                    _outputBuilder.WriteRegularField(typeNameBacking, bitfieldName);
-                    _outputBuilder.EndField(in fieldDesc);
-                }
-            }
-            else
-            {
-                currentSize = Math.Max(previousSize, currentSize);
-
-                if (_config.GenerateUnixTypes && (currentSize > previousSize))
-                {
-                    remainingBits += (currentSize - previousSize) * 8;
-                }
-
-                if (index >= 0)
-                {
-                    bitfieldName += index.ToString(CultureInfo.InvariantCulture);
-                }
-
-                var bitfieldDesc = (index > 0) ? bitfieldDescs[index - 1] : bitfieldDescs[0];
-                typeBacking = bitfieldDesc.TypeBacking;
-                typeNameBacking = GetRemappedTypeName(fieldDecl, context: null, typeBacking, out _);
-            }
-
-            var bitfieldOffset = (currentSize * 8) - remainingBits;
-            var bitwidthHexStringBacking = ((1 << fieldDecl.BitWidthValue) - 1).ToString("X", CultureInfo.InvariantCulture);
-
-            if (!IsType<BuiltinType>(fieldDecl, typeBacking, out var builtinTypeBacking))
-            {
-                AddDiagnostic(DiagnosticLevel.Warning, $"Unsupported bitfield type: '{typeBacking.TypeClassSpelling}'. Generated bindings may be incomplete.", fieldDecl);
-                return;
-            }
-
-            var isTypeBackingSigned = false;
-
-            switch (builtinTypeBacking.Kind)
-            {
-                case CXType_Char_U:
-                case CXType_UChar:
-                case CXType_UShort:
-                case CXType_UInt:
-                {
-                    bitwidthHexStringBacking += "u";
-                    break;
-                }
-
-                case CXType_ULong:
-                {
-                    if (_config.GenerateUnixTypes)
-                    {
-                        goto default;
-                    }
-
-                    goto case CXType_UInt;
-                }
-
-                case CXType_ULongLong:
-                {
-                    if (typeNameBacking is "nuint" or "UIntPtr")
-                    {
-                        goto case CXType_UInt;
-                    }
-
-                    bitwidthHexStringBacking += "UL";
-                    break;
-                }
-
-                case CXType_Char_S:
-                case CXType_SChar:
-                case CXType_Short:
-                case CXType_Int:
-                {
-                    isTypeBackingSigned = true;
-                    break;
-                }
-
-                case CXType_Long:
-                {
-                    isTypeBackingSigned = true;
-
-                    if (_config.GenerateUnixTypes)
-                    {
-                        goto default;
-                    }
-                    break;
-                }
-
-                case CXType_LongLong:
-                {
-                    isTypeBackingSigned = true;
-
-                    if (typeNameBacking is "nint" or "IntPtr")
-                    {
-                        goto case CXType_Int;
-                    }
-
-                    bitwidthHexStringBacking += "L";
-                    break;
-                }
-
-                default:
-                {
-                    AddDiagnostic(DiagnosticLevel.Warning, $"Unsupported bitfield type: '{builtinTypeBacking.TypeClassSpelling}'. Generated bindings may be incomplete.", fieldDecl);
-                    break;
-                }
-            }
-
-            var bitwidthHexString = ((1 << fieldDecl.BitWidthValue) - 1).ToString("X", CultureInfo.InvariantCulture);
-            var type = fieldDecl.Type;
-
-            if (IsType<BuiltinType>(fieldDecl, type, out var builtinType))
-            {
-                type = builtinType;
-            }
-            else if (IsType<EnumType>(fieldDecl, type, out var enumType))
-            {
-                if (IsType<BuiltinType>(fieldDecl, enumType.Decl.IntegerType, out builtinType))
-                {
-                    type = enumType;
-                }
-                else
-                {
-                    AddDiagnostic(DiagnosticLevel.Warning, $"Unsupported bitfield type: '{enumType.Decl.IntegerType.TypeClassSpelling}'. Generated bindings may be incomplete.", fieldDecl);
-                    return;
-                }
-            }
-            else
-            {
-                AddDiagnostic(DiagnosticLevel.Warning, $"Unsupported bitfield type: '{fieldDecl.Type.TypeClassSpelling}'. Generated bindings may be incomplete.", fieldDecl);
-                return;
-            }
-
-            var isTypeSigned = false;
-
-            switch (builtinType.Kind)
-            {
-                case CXType_Char_U:
-                case CXType_UChar:
-                case CXType_UShort:
-                case CXType_UInt:
-                {
-                    bitwidthHexString += "u";
-                    break;
-                }
-
-                case CXType_ULong:
-                {
-                    if (_config.GenerateUnixTypes)
-                    {
-                        goto default;
-                    }
-
-                    goto case CXType_UInt;
-                }
-
-                case CXType_ULongLong:
-                {
-                    if (typeNameBacking is "nuint" or "UIntPtr")
-                    {
-                        goto case CXType_UInt;
-                    }
-
-                    bitwidthHexString += "UL";
-                    break;
-                }
-
-                case CXType_Char_S:
-                case CXType_SChar:
-                case CXType_Short:
-                case CXType_Int:
-                {
-                    isTypeSigned = true;
-                    break;
-                }
-
-                case CXType_Long:
-                {
-                    isTypeSigned = true;
-
-                    if (_config.GenerateUnixTypes)
-                    {
-                        goto default;
-                    }
-                    break;
-                }
-
-                case CXType_LongLong:
-                {
-                    isTypeSigned = true;
-
-                    if (typeNameBacking is "nint" or "IntPtr")
-                    {
-                        goto case CXType_Int;
-                    }
-
-                    bitwidthHexString += "L";
-                    break;
-                }
-
-                default:
-                {
-                    AddDiagnostic(DiagnosticLevel.Warning, $"Unsupported bitfield type: '{builtinType.TypeClassSpelling}'. Generated bindings may be incomplete.", fieldDecl);
-                    break;
-                }
-            }
-
-            var accessSpecifier = GetAccessSpecifier(fieldDecl, matchStar: false);
-            var name = GetRemappedCursorName(fieldDecl);
-            var escapedName = EscapeName(name);
-
-            var desc = new FieldDesc {
-                AccessSpecifier = accessSpecifier,
-                NativeTypeName = nativeTypeName,
-                EscapedName = escapedName,
-                ParentName = GetRemappedCursorName(parent),
-                Offset = null,
-                NeedsNewKeyword = false,
-                Location = fieldDecl.Location,
-                HasBody = true,
-                WriteCustomAttrs = static context => {
-                    (var fieldDecl, var generator) = ((FieldDecl, PInvokeGenerator))context;
-
-                    generator.WithAttributes(fieldDecl);
-                    generator.WithUsings(fieldDecl);
-                },
-                CustomAttrGeneratorData = (fieldDecl, this),
-            };
-
-            _outputBuilder.WriteDivider();
-            _outputBuilder.BeginField(in desc);
-            _outputBuilder.WriteRegularField(typeName, escapedName);
-            _outputBuilder.BeginBody();
-
-            var recordDeclName = GetCursorName(recordDecl);
-
-            // Small types become uint32/int32s after shifting
-            var isSmallType = fieldDecl.Type.Handle.SizeOf < 4;
-            var isSmallTypeBacking = currentSize < 4;
-
-            // Check if field/backing types match
-            var isTypeMismatch = type != builtinTypeBacking;
-
-            // Signed types are sign extended when shifted
-            var isUnsignedToSigned = !isTypeBackingSigned && isTypeSigned;
-
-            // Check if type is directly shiftable/maskable
-            // Remapped types are not guaranteed to be shiftable or maskable
-            // Enums are maskable, but not shiftable
-            var isTypeLikelyRemapped = !isTypeMismatch && (typeName != typeNameBacking);
-            var isTypeAnEnum = IsType<EnumType>(fieldDecl);
-
-            // Main cases:
-            // backing  int, current int            (value << cns) >> cns
-            // backing  int, current uint           (uint)((value >> cns) & msk)
-
-            // backing uint, current int            (int)(value << cns) >> cns
-            // backing uint, current uint           (value >> cns) & msk
-
-            // backing uint, current byte           (byte)((value >> cns) & msk)
-            // backing uint, current sbyte          (sbyte)((sbyte)(value << cns) >> cns)
-
-            // Getter
-            {
-                _outputBuilder.BeginGetter(_config.GenerateAggressiveInlining,
-                    isReadOnly: !Config.GenerateCompatibleCode);
-                var code = _outputBuilder.BeginCSharpCode();
-
-                code.WriteIndented("return ");
-
-                var needsCastToFinal = (isSmallType || isTypeMismatch || isUnsignedToSigned) && !isTypeAnEnum;
-
-                // This is to handle the "backing uint, current sbyte" case
-                var needsCastToFinalAgain = isUnsignedToSigned && isSmallType;
-                if (needsCastToFinalAgain)
-                {
-                    code.Write('(');
-                    code.BeginMarker("typeName");
-                    code.Write(typeName);
-                    code.EndMarker("typeName");
-                    code.Write(")(");
-                }
-
-                if (needsCastToFinal)
-                {
-                    code.Write('(');
-                    code.BeginMarker("typeName");
-                    code.Write(typeName);
-                    code.EndMarker("typeName");
-                    code.Write(")");
-                }
-
-                var needsCastToFinalParen = needsCastToFinal && !isUnsignedToSigned;
-                if (needsCastToFinalParen)
-                {
-                    code.Write('(');
-                }
-
-                var needsCastBeforeOp = isTypeLikelyRemapped || isTypeAnEnum;
-                if (needsCastBeforeOp)
-                {
-                    code.Write('(');
-                    code.BeginMarker("typeName");
-                    code.Write(typeName);
-                    code.EndMarker("typeName");
-                    code.Write(")(");
-                }
-
-                if (isTypeSigned)
-                {
-                    code.Write('(');
-
-                    if (!string.IsNullOrWhiteSpace(contextName))
-                    {
-                        code.BeginMarker("contextName");
-                        code.Write(contextName);
-                        code.EndMarker("contextName");
-                        code.Write('.');
-                    }
-
-                    code.BeginMarker("bitfieldName");
-                    code.Write(bitfieldName);
-                    code.EndMarker("bitfieldName");
-
-                    code.Write(" << ");
-                    code.BeginMarker("remainingBitsMinusBitWidth");
-                    code.Write(remainingBits - fieldDecl.BitWidthValue);
-                    code.EndMarker("remainingBitsMinusBitWidth");
-
-                    code.Write(')');
-
-                    code.Write(" >> ");
-                    code.BeginMarker("currentSizeMinusBitWidth");
-                    code.Write((currentSize * 8) - fieldDecl.BitWidthValue);
-                    code.EndMarker("currentSizeMinusBitWidth");
-                }
-                else
-                {
-                    var needsOffset = bitfieldOffset != 0;
-                    if (needsOffset)
-                    {
-                        code.Write('(');
-                    }
-
-                    if (!string.IsNullOrWhiteSpace(contextName))
-                    {
-                        code.BeginMarker("contextName");
-                        code.Write(contextName);
-                        code.EndMarker("contextName");
-                        code.Write('.');
-                    }
-
-                    code.BeginMarker("bitfieldName");
-                    code.Write(bitfieldName);
-                    code.EndMarker("bitfieldName");
-
-                    if (needsOffset)
-                    {
-                        code.Write(" >> ");
-                        code.BeginMarker("bitfieldOffset");
-                        code.Write(bitfieldOffset);
-                        code.EndMarker("bitfieldOffset");
-
-                        code.Write(')');
-                    }
-
-                    code.Write(" & 0x");
-                    code.BeginMarker("bitwidthHexStringBacking");
-                    code.Write(bitwidthHexStringBacking);
-                    code.EndMarker("bitwidthHexStringBacking");
-                }
-
-                if (needsCastBeforeOp)
-                {
-                    code.Write(')');
-                }
-
-                if (needsCastToFinalParen)
-                {
-                    code.Write(')');
-                }
-
-                if (needsCastToFinalAgain)
-                {
-                    code.Write(')');
-                }
-
-                code.WriteSemicolon();
-                code.WriteNewline();
-                _outputBuilder.EndCSharpCode(code);
-                _outputBuilder.EndGetter();
-            }
-
-            // Setter
-            {
-                _outputBuilder.BeginSetter(_config.GenerateAggressiveInlining);
-                var code = _outputBuilder.BeginCSharpCode();
-                code.WriteIndentation();
-
-                if (!string.IsNullOrWhiteSpace(contextName))
-                {
-                    code.BeginMarker("contextName");
-                    code.Write(contextName);
-                    code.EndMarker("contextName");
-                    code.Write('.');
-                }
-
-                code.BeginMarker("bitfieldName");
-                code.Write(bitfieldName);
-                code.EndMarker("bitfieldName");
-
-                code.Write(" = ");
-
-                var needsCastToFinal = isSmallTypeBacking;
-                if (needsCastToFinal)
-                {
-                    code.Write('(');
-                    code.BeginMarker("typeNameBacking");
-                    code.Write(typeNameBacking);
-                    code.EndMarker("typeNameBacking");
-                    code.Write(")(");
-                }
-
-                // Zero out target bits
-                code.Write('(');
-
-                if (!string.IsNullOrWhiteSpace(contextName))
-                {
-                    code.Write(contextName);
-                    code.Write('.');
-                }
-
-                code.Write(bitfieldName);
-
-                code.Write(" & ~");
-
-                if (bitfieldOffset != 0)
-                {
-                    code.Write('(');
-                }
-
-                code.Write("0x");
-                code.BeginMarker("bitwidthHexStringBacking");
-                code.Write(bitwidthHexStringBacking);
-                code.EndMarker("bitwidthHexStringBacking");
-
-                if (bitfieldOffset != 0)
-                {
-                    code.Write(" << ");
-                    code.BeginMarker("bitfieldOffset");
-                    code.Write(bitfieldOffset);
-                    code.EndMarker("bitfieldOffset");
-                    code.Write(')');
-                }
-
-                // Write to target bits
-                code.Write(") | ");
-
-                var needsCastBeforeLogicalOr = isTypeMismatch && !isTypeAnEnum;
-                if (needsCastBeforeLogicalOr)
-                {
-                    code.Write('(');
-                    code.Write(typeNameBacking);
-                    code.Write(")");
-                }
-
-                code.Write('(');
-
-                if (bitfieldOffset != 0)
-                {
-                    code.Write('(');
-                }
-
-                var needsCastBeforeOp = isTypeLikelyRemapped || isTypeAnEnum;
-                if (needsCastBeforeOp)
-                {
-                    code.Write('(');
-                    code.Write(typeNameBacking);
-                    code.Write(")(value)");
-                }
-                else
-                {
-                    code.Write("value");
-                }
-
-                code.Write(" & 0x");
-                code.BeginMarker("bitwidthHexString");
-                code.Write(bitwidthHexString);
-                code.EndMarker("bitwidthHexString");
-
-                if (bitfieldOffset != 0)
-                {
-                    code.Write(") << ");
-                    code.Write(bitfieldOffset);
-                }
-
-                code.Write(')');
-
-                if (needsCastToFinal)
-                {
-                    code.Write(')');
-                }
-
-                code.WriteSemicolon();
-                code.WriteNewline();
-                _outputBuilder.EndCSharpCode(code);
-                _outputBuilder.EndSetter();
-            }
-
-            _outputBuilder.EndBody();
-            _outputBuilder.EndField(in desc);
-            _outputBuilder.WriteDivider();
-
-            remainingBits -= fieldDecl.BitWidthValue;
-            previousSize = Math.Max(previousSize, currentSize);
-        }
-
-        void VisitConstantOrIncompleteArrayFieldDecl(RecordDecl recordDecl, FieldDecl constantOrIncompleteArray)
-        {
-            if (!IsTypeConstantOrIncompleteArray(constantOrIncompleteArray, out var arrayType))
-            {
-                AddDiagnostic(DiagnosticLevel.Error, "Expected constant or incomplete array. Generated bindings may be incomplete", constantOrIncompleteArray);
-                return;
-            }
-
-            var outputBuilder = _outputBuilder;
-            var arrayTypeName = GetRemappedTypeName(constantOrIncompleteArray, context: null, arrayType, out _);
-
-            if (IsSupportedFixedSizedBufferType(arrayTypeName))
-            {
-                return;
-            }
-
-            _outputBuilder.WriteDivider();
-
-            var alignment = Math.Max(recordDecl.TypeForDecl.Handle.AlignOf, 1);
-            var maxAlignm = recordDecl.Fields.Any() ? recordDecl.Fields.Max((fieldDecl) => Math.Max(fieldDecl.Type.Handle.AlignOf, 1)) : alignment;
-
-            var accessSpecifier = GetAccessSpecifier(constantOrIncompleteArray, matchStar: false);
-            var elementType = arrayType.ElementType;
-            var isUnsafeElementType = IsTypePointerOrReference(constantOrIncompleteArray, arrayType.ElementType) &&
-                                      !arrayTypeName.Equals("IntPtr", StringComparison.Ordinal) && !arrayTypeName.Equals("UIntPtr", StringComparison.Ordinal);
-
-            var name = GetArtificialFixedSizedBufferName(constantOrIncompleteArray);
-            var escapedName = EscapeName(name);
-
-            var arraySize = Math.Max((arrayType as ConstantArrayType)?.Size ?? 0, 1);
-            var totalSize = arraySize;
-            var totalSizeString = $"{arraySize}";
-            var sizePerDimension = new List<(long index, long size)>() { (0, arraySize) };
-
-            while (IsTypeConstantOrIncompleteArray(recordDecl, elementType, out var subArrayType))
-            {
-                var subArraySize = Math.Max((subArrayType as ConstantArrayType)?.Size ?? 0, 1);
-                totalSize *= subArraySize;
-                totalSizeString += $" * {subArraySize}";
-                sizePerDimension.Add((0, subArraySize));
-
-                elementType = subArrayType.ElementType;
-            }
-
-            long alignment32 = -1;
-            long alignment64 = -1;
-
-            GetTypeSize(constantOrIncompleteArray, constantOrIncompleteArray.Type, ref alignment32, ref alignment64, out var size32, out var size64);
-
-            if ((size32 == 0 || size64 == 0) && _testOutputBuilder != null)
-            {
-                AddDiagnostic(DiagnosticLevel.Info, $"{escapedName} (constant array field) has a size of 0", constantOrIncompleteArray);
-            }
-
-            if (_config.GenerateCompatibleCode || (totalSize <= 1) || isUnsafeElementType)
-            {
-                totalSizeString = null;
-            }
-
-            var desc = new StructDesc {
-                AccessSpecifier = accessSpecifier,
-                EscapedName = escapedName,
-                IsUnsafe = isUnsafeElementType,
-                Layout = new() {
-                    Alignment32 = alignment32,
-                    Alignment64 = alignment64,
-                    Size32 = size32,
-                    Size64 = size64,
-                    Pack = alignment < maxAlignm ? alignment.ToString(CultureInfo.InvariantCulture) : null,
-                    MaxFieldAlignment = maxAlignm,
-                    Kind = LayoutKind.Sequential
-                },
-                Location = constantOrIncompleteArray.Location,
-                IsNested = true,
-                WriteCustomAttrs = static context => {
-                    (var fieldDecl, var outputBuilder, var generator, var totalSizeString) = ((FieldDecl, IOutputBuilder, PInvokeGenerator, string?))context;
-
-                    generator.WithAttributes(fieldDecl);
-                    generator.WithUsings(fieldDecl);
-
-                    if (totalSizeString is not null)
-                    {
-                        outputBuilder.WriteCustomAttribute($"InlineArray({totalSizeString})");
-                    }
-                },
-                CustomAttrGeneratorData = (constantOrIncompleteArray, _outputBuilder, this, totalSizeString),
-            };
-
-            _outputBuilder.BeginStruct(in desc);
-
-            var firstFieldName = "";
-            var numFieldsToEmit = (totalSizeString is not null) ? Math.Min(totalSize, 1) : totalSize;
-
-            for (long i = 0; i < numFieldsToEmit; i++)
-            {
-                var dimension = sizePerDimension[0];
-                var firstDimension = dimension.index++;
-                var fieldName = $"e{firstDimension}";
-                sizePerDimension[0] = dimension;
-
-                var separateStride = false;
-                for (var d = 1; d < sizePerDimension.Count; d++)
-                {
-                    dimension = sizePerDimension[d];
-                    fieldName += $"_{dimension.index}";
-                    sizePerDimension[d] = dimension;
-
-                    var previousDimension = sizePerDimension[d - 1];
-
-                    if (previousDimension.index == previousDimension.size)
-                    {
-                        previousDimension.index = 0;
-                        dimension.index++;
-                        sizePerDimension[d - 1] = previousDimension;
-                        separateStride = true;
-                    }
-
-                    sizePerDimension[d] = dimension;
-                }
-
-                if (string.IsNullOrEmpty(firstFieldName))
-                {
-                    firstFieldName = fieldName;
-                }
-
-                var fieldDesc = new FieldDesc {
-                    AccessSpecifier = accessSpecifier,
-                    NativeTypeName = null,
-                    EscapedName = fieldName,
-                    Offset = null,
-                    NeedsNewKeyword = false,
-                    Location = constantOrIncompleteArray.Location,
-                };
-
-                _outputBuilder.BeginField(in fieldDesc);
-                _outputBuilder.WriteRegularField(arrayTypeName, fieldName);
-                _outputBuilder.EndField(in fieldDesc);
-                if (!separateStride)
-                {
-                    _outputBuilder.SuppressDivider();
-                }
-            }
-
-            var generateCompatibleCode = _config.GenerateCompatibleCode;
-
-            if (generateCompatibleCode || isUnsafeElementType)
-            {
-                _outputBuilder.BeginIndexer(AccessSpecifier.Public, isUnsafe: generateCompatibleCode && !isUnsafeElementType, needsUnscopedRef: false);
-                _outputBuilder.WriteIndexer($"ref {arrayTypeName}");
-                _outputBuilder.BeginIndexerParameters();
-                var param = new ParameterDesc {
-                    Name = "index",
-                    Type = "int",
-                };
-                _outputBuilder.BeginParameter(in param);
-                _outputBuilder.EndParameter(in param);
-                _outputBuilder.EndIndexerParameters();
-                _outputBuilder.BeginBody();
-
-                _outputBuilder.BeginGetter(_config.GenerateAggressiveInlining, isReadOnly: false);
-                var code = _outputBuilder.BeginCSharpCode();
-
-                code.WriteIndented("fixed (");
-                code.Write(arrayTypeName);
-                code.Write("* pThis = &");
-                code.Write(firstFieldName);
-                code.WriteLine(')');
-                code.WriteBlockStart();
-                code.WriteIndented("return ref pThis[index]");
-                code.WriteSemicolon();
-                code.WriteNewline();
-                code.WriteBlockEnd();
-                _outputBuilder.EndCSharpCode(code);
-
-                _outputBuilder.EndGetter();
-                _outputBuilder.EndBody();
-                _outputBuilder.EndIndexer();
-            }
-            else if (totalSizeString is null)
-            {
-                _outputBuilder.BeginIndexer(AccessSpecifier.Public, isUnsafe: false, needsUnscopedRef: !_config.GenerateCompatibleCode);
-                _outputBuilder.WriteIndexer($"ref {arrayTypeName}");
-                _outputBuilder.BeginIndexerParameters();
-                var param = new ParameterDesc {
-                    Name = "index",
-                    Type = "int",
-                };
-                _outputBuilder.BeginParameter(in param);
-                _outputBuilder.EndParameter(in param);
-                _outputBuilder.EndIndexerParameters();
-                _outputBuilder.BeginBody();
-
-                _outputBuilder.BeginGetter(_config.GenerateAggressiveInlining, isReadOnly: false);
-                var code = _outputBuilder.BeginCSharpCode();
-                code.AddUsingDirective("System");
-                code.AddUsingDirective("System.Runtime.InteropServices");
-
-                code.WriteIndented("return ref ");
-
-                if (arraySize == 1)
-                {
-                    code.AddUsingDirective("System.Runtime.CompilerServices");
-                    code.Write("Unsafe.Add(ref e0, index)");
-                }
-                else
-                {
-                    code.Write("AsSpan()[index]");
-                }
-
-                code.WriteSemicolon();
-                code.WriteNewline();
-                _outputBuilder.EndCSharpCode(code);
-
-                _outputBuilder.EndGetter();
-                _outputBuilder.EndBody();
-                _outputBuilder.EndIndexer();
-
-                var function = new FunctionOrDelegateDesc {
-                    AccessSpecifier = AccessSpecifier.Public,
-                    EscapedName = "AsSpan",
-                    IsAggressivelyInlined = _config.GenerateAggressiveInlining,
-                    IsStatic = false,
-                    IsMemberFunction = true,
-                    ReturnType = $"Span<{arrayTypeName}>",
-                    Location = constantOrIncompleteArray.Location,
-                    HasBody = true,
-                    NeedsUnscopedRef = !_config.GenerateCompatibleCode,
-                };
-
-                var isUnsafe = false;
-                _outputBuilder.BeginFunctionOrDelegate(in function, ref isUnsafe);
-
-                _outputBuilder.BeginFunctionInnerPrototype(in function);
-
-                if (arraySize == 1)
-                {
-                    param = new ParameterDesc {
-                        Name = "length",
-                        Type = "int",
-                    };
-
-                    _outputBuilder.BeginParameter(in param);
-                    _outputBuilder.EndParameter(in param);
-                }
-
-                _outputBuilder.EndFunctionInnerPrototype(in function);
-                _outputBuilder.BeginBody(true);
-                code = _outputBuilder.BeginCSharpCode();
-
-                code.Write("MemoryMarshal.CreateSpan(ref ");
-                code.Write(firstFieldName);
-                code.Write(", ");
-
-                if (arraySize == 1)
-                {
-                    code.Write("length");
-                }
-                else
-                {
-                    code.Write(totalSize);
-                }
-
-                code.Write(')');
-                code.WriteSemicolon();
-                _outputBuilder.EndBody(true);
-                _outputBuilder.EndCSharpCode(code);
-                _outputBuilder.EndFunctionOrDelegate(in function);
-            }
-
-            _outputBuilder.EndStruct(in desc);
-        }
-
-        void StartTestOutput(ref bool testOutputStarted, bool generateTestsClass, bool isTopLevelStruct)
-        {
-            if ((_testOutputBuilder is not null) && generateTestsClass && !isTopLevelStruct && !testOutputStarted)
-            {
-                _testOutputBuilder.WriteIndented("/// <summary>Provides validation of the <see cref=\"");
-                _testOutputBuilder.Write(escapedName);
-                _testOutputBuilder.WriteLine("\" /> struct.</summary>");
-
-                WithAttributes(recordDecl, onlySupportedOSPlatform: true, isTestOutput: true);
-
-                _testOutputBuilder.WriteIndented("public static unsafe partial class ");
-                _testOutputBuilder.Write(escapedName);
-                _testOutputBuilder.WriteLine("Tests");
-                _testOutputBuilder.WriteBlockStart();
-
-                testOutputStarted = true;
-            }
-        }
-    }
-
     private void VisitTranslationUnitDecl(TranslationUnitDecl translationUnitDecl)
     {
         Visit(translationUnitDecl.Decls);
@@ -3449,297 +1485,6 @@ public partial class PInvokeGenerator
     private static void VisitUsingShadowDecl(UsingShadowDecl usingShadowDecl)
     {
         // Nothing to handle for binding generation
-    }
-
-    private void VisitVarDecl(VarDecl varDecl)
-    {
-        if (IsPrevContextStmt<DeclStmt>(out var declStmt, out _))
-        {
-            ForDeclStmt(varDecl, declStmt);
-        }
-        else if (IsPrevContextDecl<TranslationUnitDecl>(out _, out _) || IsPrevContextDecl<LinkageSpecDecl>(out _, out _) || IsPrevContextDecl<NamespaceDecl>(out _, out _) || IsPrevContextDecl<RecordDecl>(out _, out _))
-        {
-            if (!varDecl.HasInit)
-            {
-                // Nothing to do if a top level const declaration doesn't have an initializer
-                return;
-            }
-
-            var type = varDecl.Type;
-            var isMacroDefinitionRecord = false;
-
-            var nativeName = GetCursorName(varDecl);
-            if (nativeName.StartsWith("ClangSharpMacro_", StringComparison.Ordinal))
-            {
-                type = varDecl.Init.Type;
-                nativeName = nativeName["ClangSharpMacro_".Length..];
-                isMacroDefinitionRecord = true;
-            }
-
-            var accessSpecifier = GetAccessSpecifier(varDecl, matchStar: false);
-            var name = GetRemappedName(nativeName, varDecl, tryRemapOperatorName: false, out var wasRemapped, skipUsing: true);
-            var escapedName = EscapeName(name);
-
-            if (isMacroDefinitionRecord)
-            {
-                if (IsStmtAsWritten<DeclRefExpr>(varDecl.Init, out var declRefExpr, removeParens: true))
-                {
-                    if ((declRefExpr.Decl is NamedDecl namedDecl) && (name == GetCursorName(namedDecl)))
-                    {
-                        return;
-                    }
-                }
-                else if (IsStmtAsWritten<RecoveryExpr>(varDecl.Init, out var recoveryExpr, removeParens: true))
-                {
-                    return;
-                }
-            }
-
-            var openedOutputBuilder = false;
-            var className = GetClass(name);
-
-            if (_outputBuilder is null)
-            {
-                openedOutputBuilder = true;
-
-                if (IsUnsafe(varDecl, type) && (!varDecl.HasInit || !IsStmtAsWritten<StringLiteral>(varDecl.Init, out _, removeParens: true)))
-                {
-                    _topLevelClassIsUnsafe[className] = true;
-                }
-            }
-
-            var typeName = GetTargetTypeName(varDecl, out var nativeTypeName);
-
-            if (isMacroDefinitionRecord)
-            {
-                var nativeTypeNameBuilder = new StringBuilder("#define");
-
-                _ = nativeTypeNameBuilder.Append(' ');
-                _ = nativeTypeNameBuilder.Append(nativeName);
-                _ = nativeTypeNameBuilder.Append(' ');
-
-                var macroValue = GetSourceRangeContents(varDecl.TranslationUnit.Handle, varDecl.Init.Extent);
-                _ = nativeTypeNameBuilder.Append(macroValue.Replace(@"\\", "\\", StringComparison.Ordinal));
-
-                nativeTypeName = nativeTypeNameBuilder.ToString();
-            }
-
-            var kind = ValueKind.Unknown;
-            var flags = ValueFlags.None;
-
-            if (varDecl.HasInit)
-            {
-                flags |= ValueFlags.Initializer;
-            }
-
-            if (type.IsLocalConstQualified || isMacroDefinitionRecord || IsTypeConstantOrIncompleteArray(varDecl))
-            {
-                flags |= ValueFlags.Constant;
-            }
-
-            if (IsStmtAsWritten<StringLiteral>(varDecl.Init, out var stringLiteral, removeParens: true))
-            {
-                kind = ValueKind.String;
-
-                switch (stringLiteral.Kind)
-                {
-                    case CX_SLK_Ordinary:
-                    case CX_SLK_UTF8:
-                    {
-                        typeName = (flags & ValueFlags.Constant) != 0 ? "ReadOnlySpan<byte>" : "byte[]";
-                        break;
-                    }
-
-                    case CX_SLK_Wide:
-                    {
-                        if (_config.GenerateUnixTypes)
-                        {
-                            goto case CX_SLK_UTF32;
-                        }
-                        else
-                        {
-                            goto case CX_SLK_UTF16;
-                        }
-                    }
-
-                    case CX_SLK_UTF16:
-                    {
-                        kind = ValueKind.Primitive;
-                        typeName = "string";
-                        break;
-                    }
-
-                    case CX_SLK_UTF32:
-                    {
-                        typeName = (!_config.GenerateCompatibleCode && (flags & ValueFlags.Constant) != 0) ? "ReadOnlySpan<uint>" : "uint[]";
-                        break;
-                    }
-
-                    default:
-                    {
-                        AddDiagnostic(DiagnosticLevel.Error, $"Unsupported string literal kind: '{stringLiteral.Kind}'. Generated bindings may be incomplete.", stringLiteral);
-                        break;
-                    }
-                }
-            }
-            else if (IsPrimitiveValue(varDecl, type))
-            {
-                kind = ValueKind.Primitive;
-
-                if ((flags & ValueFlags.Constant) != 0 && !IsConstant(typeName, varDecl.Init))
-                {
-                    flags |= ValueFlags.Copy;
-                }
-                else if (_config.WithTransparentStructs.TryGetValue(typeName, out var transparentStruct))
-                {
-                    typeName = transparentStruct.Name;
-                }
-            }
-            else if ((varDecl.StorageClass == CX_SC_Static) || openedOutputBuilder)
-            {
-                kind = ValueKind.Unmanaged;
-
-                if (varDecl.HasInit)
-                {
-                    if ((varDecl.Init is CXXConstructExpr cxxConstructExpr) && cxxConstructExpr.Constructor.IsCopyConstructor)
-                    {
-                        if (cxxConstructExpr.Args[0] is CXXUuidofExpr)
-                        {
-                            // It's easiest just to let _uuidsToGenerate handle it
-                            return;
-                        }
-                        flags |= ValueFlags.Copy;
-                    }
-                }
-
-                if (IsType<ArrayType>(varDecl, type, out var arrayType))
-                {
-                    flags |= ValueFlags.Array;
-
-                    if (!_config.GenerateUnmanagedConstants)
-                    {
-                        do
-                        {
-                            typeName += "[]";
-                            arrayType = arrayType.ElementType as ArrayType;
-                        }
-                        while (arrayType is not null);
-                    }
-                }
-            }
-
-            if (typeName.Equals("Guid", StringComparison.Ordinal))
-            {
-                _ = _generatedUuids.Add(name);
-            }
-
-            var desc = new ValueDesc {
-                AccessSpecifier = accessSpecifier,
-                TypeName = typeName,
-                EscapedName = escapedName,
-                NativeTypeName = nativeTypeName,
-                Kind = kind,
-                Flags = flags,
-                Location = varDecl.Location,
-                WriteCustomAttrs = static context => {
-                    (var varDecl, var generator) = ((VarDecl, PInvokeGenerator))context;
-
-                    generator.WithAttributes(varDecl);
-                    generator.WithUsings(varDecl);
-                },
-                CustomAttrGeneratorData = (varDecl, this),
-            };
-
-            if (openedOutputBuilder)
-            {
-                StartUsingOutputBuilder(className);
-                Debug.Assert(_outputBuilder is not null);
-
-                if ((kind == ValueKind.String) && typeName.StartsWith("ReadOnlySpan<", StringComparison.Ordinal))
-                {
-                    _outputBuilder.EmitSystemSupport();
-                }
-            }
-
-            Debug.Assert(_outputBuilder is not null);
-
-            _outputBuilder.BeginValue(in desc);
-
-            var currentContext = _context.Last;
-            Debug.Assert(currentContext is not null);
-            currentContext.Value = (currentContext.Value.Cursor, desc);
-
-            if (varDecl.HasInit)
-            {
-                var dereference = IsType<PointerType>(varDecl, type, out var pointerType) &&
-                                  IsType<FunctionType>(varDecl, pointerType.PointeeType) &&
-                                  isMacroDefinitionRecord;
-
-                if (dereference)
-                {
-                    _outputBuilder.BeginDereference();
-                }
-
-                Visit(varDecl.Init);
-
-                if (dereference)
-                {
-                    _outputBuilder.EndDereference();
-                }
-            }
-
-            _outputBuilder.EndValue(in desc);
-
-            if (openedOutputBuilder)
-            {
-                StopUsingOutputBuilder();
-            }
-            else
-            {
-                _outputBuilder.WriteDivider();
-            }
-        }
-        else if (IsPrevContextDecl<FunctionDecl>(out _, out _))
-        {
-            // This should be handled in the function body as part of a DeclStmt
-        }
-        else
-        {
-            _ = IsPrevContextDecl<Decl>(out var previousContext, out _);
-            AddDiagnostic(DiagnosticLevel.Error, $"Unsupported variable declaration parent: '{previousContext?.CursorKindSpelling}'. Generated bindings may be incomplete.", previousContext);
-        }
-
-        void ForDeclStmt(VarDecl varDecl, DeclStmt declStmt)
-        {
-            var outputBuilder = StartCSharpCode();
-            var name = GetRemappedCursorName(varDecl);
-            var escapedName = EscapeName(name);
-
-            if (varDecl == declStmt.Decls[0])
-            {
-                var type = varDecl.Type;
-                var typeName = GetRemappedTypeName(varDecl, context: null, type, out _);
-
-                outputBuilder.Write(typeName);
-
-                if (IsType<ArrayType>(varDecl, type))
-                {
-                    outputBuilder.Write("[]");
-                }
-
-                outputBuilder.Write(' ');
-            }
-
-            outputBuilder.Write(escapedName);
-
-            if (varDecl.HasInit)
-            {
-                outputBuilder.Write(" = ");
-                Visit(varDecl.Init);
-            }
-
-            StopCSharpCode();
-        }
     }
 
     private bool IsConstant(string targetTypeName, Expr initExpr)
